requirements.txt
Django==2.1.5
django-admin-interface==0.9.1
django-ckeditor==5.6.1
django-colorfield==0.1.15
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-js-asset==1.1.0
Pillow==5.4.1
pytz==2018.9
sorl-thumbnail==12.5.0
Unidecode==1.0.23
